requirements.txt
streamlit
groq
pypdf2
pandas
python-dotenv
